OAK-10660: DocumentNodeStore: avoid repeated commits of :childOrder in branch commits
Added support for child order cleanup feature to DocumentMemoryFixture.
Moved tests to more appropriate test class.
Added test case for toggling the child order cleanup feature.
