+++ oak-jcr/src/test/java/org/apache/jackrabbit/oak/jcr/ManyChildrenIT.java
@@ -24,7 +24,6 @@ import javax.jcr.NodeIterator;
 import javax.jcr.Session;
 
 import org.apache.jackrabbit.oak.fixture.NodeStoreFixture;
-import org.junit.Ignore;
 import org.junit.Test;
 
 /**
@@ -77,65 +76,5 @@ public class ManyChildrenIT extends AbstractRepositoryTest {
         test.addNode("node-x");
         writer.save();
         assertTrue(test.hasNode("node-x"));
-    }
-
-    @Test
-    // @Ignore //OAK-10646
-    public void orderableAddManyChildrenWithSave() throws Exception {
-        int childCount = 1000;
-        StringBuilder prefix = new StringBuilder("");
-        for (int k = 0; k < 90; k++) {
-            prefix.append("0123456789");
-        }
-        Session session = getAdminSession();
-        Node test = session.getRootNode().addNode("test", "nt:unstructured");
-        session.save();
-        for (int k = 0; k < childCount; k++) {
-            test.addNode(prefix.toString() + k, "nt:unstructured");
-        }
-    }
-
-    @Test
-    // @Ignore //OAK-10646
-    public void moveOrderableWithManyChildren() throws Exception {
-        int childCount = 1000;
-        StringBuilder prefix = new StringBuilder("");
-        for (int k = 0; k < 90; k++) {
-            prefix.append("0123456789");
-        }
-        Session session = getAdminSession();
-        Node test = session.getRootNode().addNode("test-0", "nt:unstructured");
-        session.save();
-        for (int k = 0; k < childCount; k++) {
-            test.addNode(prefix.toString() + k, "nt:unstructured");
-            if (k % 100 == 0) {
-                session.save();
-            }
-        }
-        session.save();
-        session.move("/test-0", "/test-1");
-        session.save();
-    }
-
-    @Test
-    // @Ignore //OAK-10646
-    public void copyOrderableWithManyChildren() throws Exception {
-        int childCount = 1000;
-        StringBuilder prefix = new StringBuilder("");
-        for (int k = 0; k < 90; k++) {
-            prefix.append("0123456789");
-        }
-        Session session = getAdminSession();
-        Node test = session.getRootNode().addNode("test-0", "nt:unstructured");
-        session.save();
-        for (int k = 0; k < childCount; k++) {
-            test.addNode(prefix.toString() + k, "nt:unstructured");
-            if (k % 100 == 0) {
-                session.save();
-            }
-        }
-        session.save();
-        session.getWorkspace().copy("/test-0", "/test-1");
-        session.save();
     }
 }
