+++ oak-store-document/src/test/java/org/apache/jackrabbit/oak/fixture/DocumentMemoryFixture.java
@@ -22,12 +22,15 @@ package org.apache.jackrabbit.oak.fixture;
 import org.apache.jackrabbit.oak.plugins.document.DocumentMK;
 import org.apache.jackrabbit.oak.plugins.document.DocumentNodeStore;
 import org.apache.jackrabbit.oak.spi.state.NodeStore;
+import org.apache.jackrabbit.oak.spi.toggle.Feature;
 
 public class DocumentMemoryFixture extends NodeStoreFixture {
 
     @Override
     public NodeStore createNodeStore() {
-        return new DocumentMK.Builder().getNodeStore();
+        DocumentMK.Builder builder = new DocumentMK.Builder();
+        builder.setNoChildOrderCleanupFeature(Feature.newFeature("FT_NOCOCLEANUP_OAK-10660", getWhiteboard()));
+        return builder.getNodeStore();
     }
 
     @Override
